// frontend/src/core/api/generated/model/puantajView.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kızılkan Şantiye API
 * OpenAPI spec version: 1.0
 */
import type { DayMarkView } from './dayMarkView'
import type { RosterEntryView } from './rosterEntryView'

export interface PuantajView {
  from: string
  to: string
  today: string
  entries: RosterEntryView[]
  marks: DayMarkView[]
}